Fixed an issue where HubGenerator.cs would throw an index out of range exception on build when the hub interface comes from a Class Library Created a Samples.Common Class Library
