+++ SignalR.Strong.Tests.xUnit/Basic.cs
@@ -3,6 +3,7 @@ using System.Threading.Tasks;
 using FluentAssertions;
 using Microsoft.AspNetCore.SignalR.Client;
 using Moq;
+using SignalR.Strong.Samples.Common.Hubs;
 using SignalR.Strong.Tests.Common;
 using Xunit;
