+++ SignalR.Strong.Tests.xUnit/SignalR.Strong.Tests.xUnit.csproj
@@ -23,6 +23,7 @@
   </ItemGroup>
 
   <ItemGroup>
+    <ProjectReference Include="..\SignalR.Strong.Samples.Common\SignalR.Strong.Samples.Common.csproj" />
     <ProjectReference Include="..\SignalR.Strong.Tests.Common\SignalR.Strong.Tests.Common.csproj" />
     <ProjectReference Include="..\SignalR.Strong\SignalR.Strong.csproj" />
   </ItemGroup>
